+++ tests/Pim/Product/DivisionServiceTest.php
@@ -1,28 +1,31 @@
 <?php
 
 /**
- * @copyright Copyright (c) Wizacha
- * @license Proprietary
+ * @author      Wizacha DevTeam <dev@wizacha.com>
+ * @copyright   Copyright (c) Wizacha
+ * @license     Proprietary
  */
 
-namespace Wizaplace\SDK\Tests\Division;
+declare(strict_types=1);
 
-use Wizaplace\SDK\Division\DivisionService;
+namespace Wizaplace\SDK\Tests\Pim\Product;
+
 use Wizaplace\SDK\Division\DivisionSettings;
+use Wizaplace\SDK\Pim\Product\DivisionService;
 use Wizaplace\SDK\Tests\ApiTestCase;
 
 class DivisionServiceTest extends ApiTestCase
 {
-    public function testGetDivisionsSettings(): void
+    public function testGetProductDivisionsSettings(): void
     {
         $divisionService = $this->buildDivisionService();
-        $divisionsSettings = $divisionService->getDivisionsSettings();
+        $divisionsSettings = $divisionService->getDivisionsSettings(1);
 
         static::assertEquals(['FR'], $divisionsSettings->getIncluded());
         static::assertEquals(['FR-59', 'FR-IDF'], $divisionsSettings->getExcluded());
     }
 
-    public function testPatchDivisionsSettings(): void
+    public function testPatchProductDivisionsSettings(): void
     {
         $divisionService = $this->buildDivisionService();
         $data = [
@@ -30,15 +33,15 @@ class DivisionServiceTest extends ApiTestCase
             'excluded' => ['FR-ARA', 'FR-NC']
         ];
         $divisionsSettings = new DivisionSettings($data);
-        $divisionService->patchDivisionsSettings($divisionsSettings);
+        $divisionService->patchDivisionsSettings(1, $divisionsSettings);
 
-        static::assertEquals($data['included'], $divisionService->getDivisionsSettings()->getIncluded());
-        static::assertEquals($data['excluded'], $divisionService->getDivisionsSettings()->getExcluded());
+        static::assertEquals($data['included'], $divisionService->getDivisionsSettings(1)->getIncluded());
+        static::assertEquals($data['excluded'], $divisionService->getDivisionsSettings(1)->getExcluded());
     }
 
     private function buildDivisionService(
-        string $email = 'admin@wizaplace.com',
-        string $password = 'password'
+        string $email = 'vendor@world-company.com',
+        string $password = 'password-vendor'
     ): DivisionService {
         $apiClient = $this->buildApiClient();
         $apiClient->authenticate($email, $password);